Moved ODataSystemQuery to seperate file, refactored api_urls.py
